Cast fields in a view based on type hint for datetime, date, time, timedelta, and uuid.
