+++ example/unicorn/components/test_datetime.py
@@ -1,6 +1,9 @@
 from datetime import datetime
+
 from django.utils import timezone
+
 from django_unicorn.components import UnicornView
+
 
 class TestDatetimeView(UnicornView):
     dt: datetime = None
