+++ example/unicorn/templates/unicorn/test-datetime.html
@@ -1,7 +1,7 @@
 {% load unicorn %}
 <div>
-    <div style="width:200px;height:200px;margin:10px;background-color:#aaffaa" unicorn:click="foo()">
-        {{ dt }}<br>
-        (click me)
-    </div>
+  <div style="width:200px;height:200px;margin:10px;background-color:#aaffaa" unicorn:click="foo()">
+    {{ dt }}<br>
+    (click me)
+  </div>
 </div>
